Split ddate into binary and library
